# openapi.json
{"openapi":"3.0.2","info":{"title":"Recipe service","version":"0.1.0"},"paths":{"/users/create":{"post":{"tags":["users"],"summary":"Create User","operationId":"create_user_users_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{id}/update":{"put":{"tags":["users"],"summary":"Update User","operationId":"update_user_users__id__update_put","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/users/{id}":{"get":{"tags":["users"],"summary":"Read User","operationId":"read_user_users__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["users"],"summary":"Read Users","operationId":"read_users_users__get","parameters":[{"required":false,"schema":{"title":"Limit","type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Users Users  Get","type":"array","items":{"$ref":"#/components/schemas/UserProfile"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recipes/create":{"post":{"tags":["recipes"],"summary":"Create Recipe","operationId":"create_recipe_recipes_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/recipes/{id}":{"get":{"tags":["recipes"],"summary":"Read Recipe","operationId":"read_recipe_recipes__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recipes/":{"get":{"tags":["recipes"],"summary":"Read Recipes","operationId":"read_recipes_recipes__get","parameters":[{"required":false,"schema":{"title":"Filter","type":"string"},"name":"filter","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Recipes Recipes  Get","type":"array","items":{"$ref":"#/components/schemas/Recipe"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recipes/{id}/update":{"put":{"tags":["recipes"],"summary":"Update Recipe","operationId":"update_recipe_recipes__id__update_put","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/recipes/{id}/delete":{"delete":{"tags":["recipes"],"summary":"Delete Recipe","operationId":"delete_recipe_recipes__id__delete_delete","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/admin/users/{id}/update":{"put":{"tags":["admin"],"summary":"Update User","operationId":"update_user_admin_users__id__update_put","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Admin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/recipes/{id}/update":{"put":{"tags":["admin"],"summary":"Update Recipe","operationId":"update_recipe_admin_recipes__id__update_put","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Admin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Admin":{"title":"Admin","required":["token","is_active"],"type":"object","properties":{"token":{"title":"Token","type":"string"},"is_active":{"title":"Is Active","type":"boolean"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Login":{"title":"Login","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}}},"Recipe":{"title":"Recipe","required":["name","dish_id","description","cooking_steps","image","user_id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"dish_id":{"title":"Dish Id","type":"integer"},"description":{"title":"Description","type":"string"},"cooking_steps":{"title":"Cooking Steps","type":"string"},"image":{"title":"Image","type":"string"},"id":{"title":"Id","type":"string"},"user_id":{"title":"User Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean"},"likes_count":{"title":"Likes Count","type":"integer"},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"}}},"RecipeIn":{"title":"RecipeIn","required":["name","dish_id","description","cooking_steps","image"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"dish_id":{"title":"Dish Id","type":"integer"},"description":{"title":"Description","type":"string"},"cooking_steps":{"title":"Cooking Steps","type":"string"},"image":{"title":"Image","type":"string"}}},"Token":{"title":"Token","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"User":{"title":"User","required":["username","hashed_password"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"username":{"title":"Username","type":"string"},"is_active":{"title":"Is Active","type":"boolean"},"favorites_count":{"title":"Favorites Count","type":"integer"},"hashed_password":{"title":"Hashed Password","type":"string"},"created_at":{"title":"Created At","type":"string","format":"date-time"},"updated_at":{"title":"Updated At","type":"string","format":"date-time"}}},"UserIn":{"title":"UserIn","required":["username","password","password2"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","minLength":8,"type":"string"},"password2":{"title":"Password2","type":"string"}}},"UserProfile":{"title":"UserProfile","required":["id","username","is_active","recipes_count"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"username":{"title":"Username","type":"string"},"is_active":{"title":"Is Active","type":"boolean"},"recipes_count":{"title":"Recipes Count","type":"integer"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"JWTBearer":{"type":"http","scheme":"bearer"}}}}